backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Baby Diagnostics backend — FastAPI + OpenCV + Claude"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "python-multipart>=0.0.18",
    "websockets>=14.0",
    "sqlalchemy[asyncio]>=2.0.36",
    "asyncpg>=0.30.0",
    "alembic>=1.14.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.7.0",
    "python-dotenv>=1.0.1",
    "pyyaml>=6.0.2",
    "python-jose[cryptography]>=3.3.0",
    "cryptography>=44.0.0",
    "google-auth>=2.37.0",
    "opencv-python-headless>=4.9.0",
    "numpy>=2.0.0",
    "pillow>=11.0.0",
    "anthropic>=0.42.0",
    "requests>=2.32.5",
]

[dependency-groups]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.25.0",
    "httpx>=0.28.0",
    "ruff>=0.8.0",
    "coverage>=7.6.0",
]
